requirements.txt
#Python 3.9.13
ahk==1.8.0
catboost==1.2.7
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.54.0
fsspec==2024.2.0
graphviz==0.20.3
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mss==9.0.2
networkx==3.2.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sympy==1.12
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
typing_extensions==4.12.2
tzdata==2024.2
zipp==3.20.2
